restructure ignore files
