+++ .gitignore
@@ -14,7 +14,6 @@ vendor
 build
 dest
 dist
-lib
 lib-cov
 coverage
 nbproject
