--- .npmignore
@@ -1,7 +1,0 @@
-test
-.travis.yml
-benchmark
-CONTRIBUTING.md
-src
-.gitignore
-babel.config.json
